Enable sql value visitor handlers modification without specifying an entire new configuration
