+++ src/main/java/com/intuit/graphql/filter/visitors/ExpressionValueHandler.java
@@ -1,0 +1,11 @@
+package com.intuit.graphql.filter.visitors;
+
+import com.intuit.graphql.filter.ast.ExpressionValue;
+import com.intuit.graphql.filter.ast.Operator;
+
+public interface ExpressionValueHandler {
+    void handle(
+            Operator operator,
+            StringBuilder expressionBuilder,
+            ExpressionValue<? extends Comparable> expressionValue);
+}
